добавлены компоненты
