+++ project/src/components/comment/comment.tsx
@@ -7,30 +7,32 @@ type CommentProps = {
 };
 
 function Comment({ review }: CommentProps): JSX.Element {
+  const {user, rating, comment, date } = review;
+
   return (
     <li className="reviews__item">
       <div className="reviews__user user">
         <div className="reviews__avatar-wrapper user__avatar-wrapper">
           <img
             className="reviews__avatar user__avatar"
-            src={review.user.avatarUrl}
+            src={user.avatarUrl}
             width="54"
             height="54"
-            alt={`Reviews avatar ${review.user.name}`}
+            alt={`Reviews avatar ${user.name}`}
           />
         </div>
-        <span className="reviews__user-name">{review.user.name}</span>
+        <span className="reviews__user-name">{user.name}</span>
       </div>
       <div className="reviews__info">
         <div className="reviews__rating rating">
           <div className="reviews__stars rating__stars">
-            <span style={{ width: getRatingStars(review.rating) }} />
+            <span style={{ width: getRatingStars(rating) }} />
             <span className="visually-hidden">Rating</span>
           </div>
         </div>
-        <p className="reviews__text">{review.comment}</p>
-        <time className="reviews__time" dateTime="2019-04-24">
-          {dayjs(review.date).format('MMMM YYYY')}
+        <p className="reviews__text">{comment}</p>
+        <time className="reviews__time" dateTime={date}>
+          {dayjs(date).format('MMMM YYYY')}
         </time>
       </div>
     </li>
