+++ project/src/components/favorites-card/favorites-card.tsx
@@ -1,0 +1,66 @@
+import { OfferType } from '../../types/offer-type';
+import { place } from '../../const';
+import { Link } from 'react-router-dom';
+import { getRatingStars } from '../../utils';
+
+type OfferCardProps = {
+  offer: OfferType;
+};
+
+function FavoritesPlaces({offer}: OfferCardProps): JSX.Element {
+  const {previewImage, price, title, type, rating, isPremium, isFavorite, id} = offer;
+
+  return (
+    <article className="favorites__card place-card">
+      {isPremium ? <div className="place-card__mark"><span>Premium</span></div> : ''}
+      <div className="favorites__image-wrapper place-card__image-wrapper">
+        <a href="#">
+          <img
+            className="place-card__image"
+            src={previewImage}
+            width="150"
+            height="110"
+            alt="Place image"
+          />
+        </a>
+      </div>
+      <div className="favorites__card-info place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">&euro;{price}</b>
+            <span className="place-card__price-text">
+            &#47;&nbsp;night
+            </span>
+          </div>
+          <button
+            className={`place-card__bookmark-button ${isFavorite ? 'place-card__bookmark-button--active' : ''} button`}
+            type="button"
+          >
+            <svg
+              className="place-card__bookmark-icon"
+              width="18"
+              height="19"
+            >
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">In bookmarks</span>
+          </button>
+        </div>
+        <div className="place-card__rating rating">
+          <div className="place-card__stars rating__stars">
+            <span style={{width: getRatingStars(rating)}}></span>
+            <span className="visually-hidden">Rating</span>
+          </div>
+        </div>
+        <h2 className="place-card__name">
+          <Link to={`/offer/:${id}`}>
+            {title}
+          </Link>
+        </h2>
+        <p className="place-card__type">{place[type]}</p>
+      </div>
+    </article>
+  );
+}
+
+export default FavoritesPlaces;
