+++ project/src/components/loading/loading.tsx
@@ -1,0 +1,11 @@
+import './loading.css';
+
+function Loading(): JSX.Element {
+  return (
+    <div className='lds-container'>
+      <div className="lds-hourglass"></div>
+    </div>
+  );
+}
+
+export default Loading;
