+++ project/src/components/place-card/place-card.tsx
