+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,66 +1,44 @@
 import Logo from '../../components/logo/logo';
 import FavoritesList from '../../components/favorites-list/favorites-list';
 import {OfferTypes} from '../../types/offer-type';
+import Nav from '../../components/nav/nav';
 
 type FavoritesScreenProps = {
   offers: OfferTypes;
 }
 
-function FavotitesScreen(props: FavoritesScreenProps): JSX.Element {
-  const {offers} = props;
+function FavotitesScreen({offers}: FavoritesScreenProps): JSX.Element {
+  const favoriteOffers = offers.filter((offer) => offer.isFavorite);
+
+  const favoritesTitle = favoriteOffers.length ? 'Saved listing' : 'Favorites (empty)';
+  const favoritesAddedBlock = favoriteOffers.length ? <FavoritesList offers={favoriteOffers} /> : (
+    <div className="favorites__status-wrapper">
+      <b className="favorites__status">Nothing yet saved.</b>
+      <p className="favorites__status-description">Save properties to narrow down search or plan your future trips.</p>
+    </div>
+  );
 
   return (
     <div className="page">
       <header className="header">
         <div className="container">
           <div className="header__wrapper">
-            <div className="header__left">
-              <Logo />
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                      Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
+            <Logo />
+            <Nav />
           </div>
         </div>
       </header>
-
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
-            <h1 className="favorites__title">Saved listing</h1>
-
-            <FavoritesList offers={offers}/>
-
+            <h1 className="favorites__title">{favoritesTitle}</h1>
+            {favoritesAddedBlock}
           </section>
         </div>
       </main>
       <footer className="footer container">
         <a className="footer__logo-link" href="main.html">
-          <img
-            className="footer__logo"
-            src="img/logo.svg"
-            alt="6 cities logo"
-            width="64"
-            height="33"
-          />
+          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
         </a>
       </footer>
     </div>
@@ -68,3 +46,4 @@ function FavotitesScreen(props: FavoritesScreenProps): JSX.Element {
 }
 
 export default FavotitesScreen;
+
