+++ project/src/pages/loading-screen/loading-screen.css
@@ -1,0 +1,44 @@
+.spin-wrapper{
+    position: relative;
+    width: 100%;
+    height: 100vh;
+    background: #fff;
+}
+.spinner{
+    position: absolute;
+    height: 60px;
+    width: 60px;
+    border: 3px solid transparent;
+    border-top-color: #A04668;
+    top: 50%;
+    left: 50%;
+    margin: -30px;
+    border-radius: 50%;
+    animation: spin 2s linear infinite;
+}
+.spinner::before, .spinner::after{
+    content:'';
+    position: absolute;
+    border: 3px solid transparent;
+    border-radius: 50%;
+}
+.spinner::before{
+    border-top-color: #254E70;
+    top: -12px;
+    left: -12px;
+    right: -12px;
+    bottom: -12px;
+    animation: spin 3s linear infinite;
+}
+.spinner::after{
+    border-top-color: #6ef500;
+    top: 6px;
+    left: 6px;
+    right: 6px;
+    bottom: 6px;  
+    animation: spin 0.5s linear infinite;
+}
+@keyframes spin{
+0% {transform: rotate(0deg);}
+100% {transform: rotate(360deg);}
+}
