+++ project/src/pages/page-404/page-404.tsx
@@ -1,4 +1,5 @@
 import {Link} from 'react-router-dom';
+import './not-found-screen.css';
 
 function Page404(): JSX.Element {
   return (
